:recycle: Separate path processes from `note.ts`
